Start of public codebase
This version of the website is intended to exactly represent the code
that is running on http://www.kickback-kingdom.com/

However, in its current state, it will probably not work if anyone
tries to download it and run it in a local HTTP server (e.g.: apache).
That's because the version in this commit has been stripped of
all image assets, because they were quite large space-wise, and
we'd like to start the git repo off without that in the git history.

Even more importantly, the website probably won't work without
a database that is populated with the correct records. This repo
DOES contain the MySQL schema used by the website, but it may
be a while before we can create a test database (with working
records and a faux admin access) that can allow the website to
work on other peoples' test system without leaking any potentially
sensitive data from the website.

This version of the codebase is still very important because it
allows anyone without admin access to see how scripts work, report bugs,
create pull requests, and so on.