Contents of beta / main
The previous commit represents what was in the website `prod`
or "production" code, i.e. that what was running the main site.

This commit represents all the in-development changes that have
been made to the website but not yet released. For anyone familiar
with the "beta" feature in the website: this commit brings the
`main` branch (the development branch) up to date with the
website's `beta` features.