Move contents of `html-prod-only` folder into `html` folder.
This makes the code in the repository better reflect what things
look like in the actual server's `/var/www/html` directory.

These things were split out into a separate folder before
because placing them all in `/html` would have caused the
site to carry two copies of the `html/assets/media` and
`html/assets/images` folders: one for the root `html`
directory, and a copy for the `html/beta` subdirectory
(that is populated by the `makefile` install script).
So it would generate a website that takes 600MB, instead
of taking 300MB with the assets (and a few other things,
for legacy reasons) factored out into their own directory.

Now that this version of the repository has all of the
image data stripped out of it, there is no longer such a
punishment for just duplicating the website's code
when releasing a website with both `prod` and `beta`/`main`
code that are available at the same time.

This particular commit only does the above for the `prod`
branch of the repository. There will be a corresponding
commit in the `main` branch that does the same thing,
but with a file selection that matches that branch's
files instead of this branch's files.