Image upload bugfixes and increased robusteness.
Adjusted hardcoded `/var/www/html` -> `/var/www/kickback-kingdom-prod/html`
(We'll have to move that into a config file or something,
or maybe find a way to use relative paths, but... later.
This commit is mostly just to get that working again.)

The other problem we encountered was that the permissions
of the `html/assets/media` folder did not allow the `apache`
user/group to actually create/write the images files that
were uploaded. This was fixed by changing the ownership of
that folder and its subfolders (also `html/assets/images`).

Code-wise, this change also makes it less likely that the
database and filesystem contents will fall out-of-sync if
an image upload fails. Now, the database transaction for
creating an image reference will rollback if the image file
was not successfully created on the filesystem (for any reason).

These changes were made by Chad and Alibaba collaboratively.