SECURITY: Remove Discord API key and KK service key.
These have been moved into the service-credentials.ini file so that they
will not be publicly visible in the git repository.

The repository will be scrubbed of these keys; that is, all instances of
those credentials will be removed from the repository's history.

This entails that commits prior to this one may not work, as invalid
values will be substituted into those key-containing variables retroactively.

This commit also introduces the `LoadServiceCredentialsOnce()` function
and modifies existing code to use that instead of the
`LoadServiceCredentials()` function. Both of these functions now return
the credentials array, which allows calling code to avoid using global
variables entirely, while still benefitting from caching of the
credential list in static/global variables inside the
service-credentials-ini.php script. (The global variables are still
publicly visible, for PHP reasons, but the global variables should not
be used. Just call `LoadServiceCredentialsOnce()` and use its
return value instead.)