Remove KKAPI line from `base-page-javascript.php`
Although we had already commented this line out to make sure the
`prod` branch would function correctly, the line still makes it easier
for branches to desync. So, in preparation for separating the API's JS
client (draft) from the other branches, this line has been removed from
`prod` so that things will be easier to sync up.