Remove the API's JavaScript client.
This commit is really intended to do the removal in the `main` branch,
with the understanding that it will be added back in to the
`topic-jsapi` branch in a later commit. So, this commit is part of
"transferring" those (originally not-version-controlled) changes from
the `main` branch and into the `topic-jsapi` branch.