Bugfix to make beta work again; autoloader tweaks.
Autoloader now uses the `KICKBACK_SCRIPT_ROOT` $_SERVER environment
variable whenever it is available. This is helpful for detecting which
local directory to execute scripts from, since this can change when
someone is logged into the beta version of the site (and it can change
without changing DOCUMENT_ROOT).

This also means that the  `$_SERVER["KICKBACK_SCRIPT_ROOT"]` is placed
into the `Kickback\_SCRIPT_ROOT` constant definition. If that server
environment variable isn't available, it falls back to
`$_SERVER["DOCUMENT_ROOT"]` and if not that, then `__DIR__ . "/.."`
for execution in CLI contexts.

Also replaced beta detection logic in
`base-page-pull-active-account-info.php`
with an if-statement that uses `$_SERVER["KICKBACK_IS_BETA"]`
to explicitly detect beta status.