Rename `_SCRIPT_ROOT` -> `SCRIPT_ROOT`
I'm not sure why I put the leading underscore in to begin with, though I
suspect it was to ensure that this special constant is distinct from
anything else that might crop up in the Kickback namespace.

It was also called _DOC_ROOT originally, but I realized that it's
important to name it something else, because it really isn't the
document root. It's for scripts specifically. And it can be different
from the document root, especially when the system is in `beta` mode.
So the underscore may have made more sense in that context.

Also, let's sure hope nothing else is named any variant of "SCRIPT_ROOT",
because that would be tremendously confusing.

It's also not going to collide with classnames because of the
all-uppercase (screaming-snake-case) style that constants are written
in.

So at this point I don't think the underscore is necessary.
It can look nicer :)