Remove DIRECTORY_SEPARATOR from require_once
Testing indicated that '/' works on both Linux AND Windows (tested by
running `php` from a `cmd` shell on a Windows machine), suggesting that
PHP already converts '/' to DIRECTORY_SEPARATOR on-the-fly for `require`
statements. So we don't need to write terribly verbose code in our
`require` statements; just always use '/'.

More like this:
https://stackoverflow.com/a/71791252/1261963

This also trials the `\Kickback\SCRIPT_ROOT` constant in `index.php` by
using it to locate the `verifySession.php` script.