Move `service-credentials-ini.php` into a class.
It is now `Kickback/Config/ServiceCredentials.php`.
This demonstrates autoloading and opens the possibility for more generic
config-loading code. (Because it would be nice to also handle
system-level config that isn't secret.)

This commit also updates some of the surrounding code a bit, mostly by
replacing `_SERVER['DOCUMENT_ROOT']` with `\Kickback\SCRIPT_ROOT` when
used in `require(_once)` statements.