Bugfix: Make 3rd arg of APIResponse ctor optional
This fixes errors like this:
```
PHP Fatal error:  Uncaught ArgumentCountError: Too few arguments to function APIResponse::__construct(), 2 passed in ...
```

I considered the possibility that the 3rd parameter was already optional
before the change that added type-safety to APIResponse:
commit 1135ed011b7ad706858e5a91cfb46e2c0c14fa8c

However, the code before that commit looked like it also did not provide
a default value for the 3rd argument. So I am still not sure how anything
would have made a 2-argument call work. Perhaps PHP was more permissive
about such things back when that code was written. Or maybe the 2-arg
calls just never worked in the first place, and just mostly showed up
in error handlers that didn't get called during normal operation.
(This speculation mostly doesn't matter, but it would have been a little
nice to know what caused this, as this kind of knowledge can sometimes
be used to find similar errors in the same codebase, and fix them.)

Regardless of what caused this regression, or if it's even a regression,
things should work a _little_ better now.