PHPStan config and composer setup
With this so far, my command for running phpstan looks like this:
```
./html/vendor/composer/phpstan/phpstan/phpstan analyse -v -c ./meta/phpstan.neon
```

I wish I could make that shorter, but it's not clear how.
(Not without, like, changing my PATH to include that subdirectory of
this project that may or may not remain at the same location on my
filesystem. Ewwww. There HAS to be a better way!)