Fix vendor class (ex: PHPMailer) autoloading
This makes the workaround at the top of `forgot-password.php` no longer
necessary, so that is removed in this commit.

Additionally, in the process of using PHPStan to screen the new changes
for errors, `html/Kickback/InitializationScripts/error_handling_init.php`
popped up with some PHPStan errors, and those are addressed in this
commit. Documentation was improved and TODOs were added.