Draft of DatabaseRow class.
This should provide some type-safety that is currently not provided when
the `array` type is used to store database/SQL query rows.

It also provides fail-early behavior whenever the caller attempts to
access a non-existant field on a database row (or the query returns two
or more columns with the same name).

The current implementation passes PHPStan with 0 errors, but has not
been tested with actual execution.

So this _probably_ has a few bugs in it. However, I did want to commit
it as early as possible so that everyone else can see its API and
concepts, and possibly start playing with it in other parts of the
project.