Fix "An error occurred" when accessing quests.
(Or, at least, attempt to fix it. It's hard to test.)
This error only occurred when logged in. And it makes sense when looking
at the code; because it occurs in a block that's guarded with `Kickback\Services\Session::isLoggedIn()`.
It's still not an obvious problem, because we'd expect Version:: to be
available in quest-signup.php if it was `require`'d from a file that
already executed `use Kickback\Common\Version;`. However, it's an
assumption that it'd remain imported (`use`d) across inclusion
boundaries, so maybe that's what caused the error.