Renamed `request...(Response)` functions to `query...(AsResponse)`
(Background:
These were originally `get...` functions, but they weren't getters,
and they induce network (or socket) I/O, which is something I'd never expect
from a "getter" function! So I changed some of these to `request...`
as I was fixing PHPStan errors. There are many more to go, but this is a
start.)

Now I'm changing my mind about how I named them earlier.

The `request...` name worked alright in most cases, but could sound ...
off ... in a few instances.

For example: `requestQuestRewardsByQuestId`
... could sound like we're asking to be rewarded for a quest, rather
than like we're querying for a list of existing rewards.

So it can at least sound a bit awkward when the thing being requested
isn't just a noun, but is also a verb that's requestable, or perhaps
even anything that could be "requested" in a different way.

I also think that `query` matches what these functions do better than
`request`, because we are making an SQL _query_ and not an HTTP _request_.
It is more suggestive of the correct type of network|socket I/O that
is performed within the functions.

Previous changes also split the original `get` functions into multiple
`request...` functions. There would be one version that returns a `Response`
object, and then a wrapper that returns or populates a strictly-typed
variable instead, thus allowing downstream code to be type-checked and
pass PHPStan scanning.

To keep the names distinct, the functions that return a `Response` type
would always have `Response` in their name somewhere, while the
strictly-typed version would just not be annotated with type
information (it would be redundant).

The problem was that the placement of this `Response` notation within
the names didn't always feel consistent.

For example: `requestQuestsResponseByQuestLineId`
Note that the word `Response` isn't at the end of the function, as it is
in many other cases, such as `requestMediaDirectoriesResponse`.

But now, the response-returning functions ALWAYS end in `...AsResponse`.
This seems to mostly-kinda work out grammatically and semantically as
far as English readability goes, and it's very consistent. Thus, this
commit makes it so.

The previous examples then become:
`queryQuestsByQuestLineIdAsResponse`
and `queryMediaDirectoriesAsResponse`