Create the `StaticClassTrait` to remove some boilerplate.
PHP currently doesn't have a way to define a `static class`, nor does it
allow an `abstract final` class (e.g.: a class that can't be
instantiated and can't be inherited from).

Such classes are useful for various reasons. They're a more hygienic way
to write free functions, and can be used to house constants that don't
make sense in an enum. (Also: enums are also currently kinda weak: it's
not possible to use enum members to initialize other enum members, which
makes it hard to do things like construct self-updating bitmasks and
boundary values. They just aren't good for bitwise things. Which means
that constants-in-classes are often still The Way.)

Anyhow, to define a static class, we would have to do something like
this:
```
final class MyStaticClass
{
    ... class contents, all static methods and constants ...

    // Prevent instantiation/construction of the (static/constant) class.
    /** @return never */
    private function __construct() {
        throw new \Exception("Instantiation of static class " . get_class($this) . "is not allowed.");
    }
}
```

The `final` prevents inheritence, and the private constructor that
always throws will prevent instantiation (at least, at runtime,
though PHPStan might pick up on the `@return never` hint to flag any code
that expects execution to continue after constructing the thing).

The downside is that the constructor is a bunch of boilerplate that
always has to be copy-pasted whenever a new class is made.

So I moved it into a trait.

Now the above class would look like so:
```
final class MyStaticClass
{
    ... class contents, all static methods and constants ...

    use \Kickback\Common\StaticClassTrait;
}
```

Just one extra line in the class, and we're good. It's even almost kinda
possible to memorize what needs to be written.