Add caller-finding function and `(min|max)_pair` functions.
The caller finding function is called
`Meta::outermost_caller_name_within_class()`

I'm pretty excited about that one.

It allows code to print debug messages that include the name of the
function that the caller actually called, and not the
several-stack-frames-deep function that's in the called function's
implementation stack.

The (min|max)_pair functions were kind of a distraction; they implement
some functionality in an upcoming trait (ClassOfConstantIntegersTrait).
While that trait will be pretty crucial, I'm not sure what drove me to
start implementing the (min|max)_pair functions; though it WAS insulting
that PHP didn't have a built-in way to do this, or a quick way that
doesn't involve copying or modifying the associative array
being examined (a potentially slow operation, and error-prone if it
involves modifying existing arrays). Anyhow, these exist now, and might
be helpful at some point. They come with tests and no dependencies, so
should be pretty low-burden. We may as well keep them.