autoloader: Ignore `never` symbol and symbols ending in `_a`.
So far, it seems like PHPStan is always autoloading the symbols that
appear in `@phpstan-type` declarations. These should be "local aliases",
and the declared symbols are not PHP classes (or any importable type).

It _does_ still declare the symbols, and the "local alias" feature does
work as a way to practice DRY with more complicated type hints,
but it also makes PHPStan seemingly emit a bunch of error messages about
classes that could not be autoloaded (and the names are the names of the
local aliases, prefixed with namespaces).

These errors are not from PHPStan, but instead from the autoloader.
And while they don't directly hurt anything, they do make it a lot
harder to read PHPStan output. And the problem is only getting worse as
PHPStan is used more and more heavily.

I really wish PHPStan wouldn't attempt to autoload symbols with the same
names as local aliases. But until that's resolved, we have a workaround:

Any symbol that is all lowercase (e.g. snake_case) and ends in `_a`
(and that is within a Kickback namespace) will be ignored by the
autoloader.

In a sense, we can use the names of the local aliases to manually tell
the autoloader that "this is a local alias in PHPStan, not a PHP class".
Then it won't try in vain to load the non-existant classes, and won't
emit noisy error messages for the non-problem.

This commit doesn't actually change those symbols. Perhaps another will,
in a bit.