Big autoloader feature adds! Subclasses and Trait classes.
This commit also gives us the ability to ignore/prevent
the autoloading of PHPStan local aliases:
* PHPDoc-side, we make sure the local aliases are lowercase and end in '_a'
* Autoloader-side, all symbols that are lowercase and end in '_a' are ignored.

That should get rid of a lot of PHPStan spam once some local aliases are
renamed.

As for the subclasses and trait classes:

--- Trait Classes ---
If there is a trait named Kickback\Foobar\SomethingTrait
living in the file html/Kickback/Foobar/SomethingTrait.php
then it is now possible to have a _class_ called
Kickback\Foobar\Something also live in that same file.

This makes it a lot easier to keep related code together, and prevents
interdependencies from closely coupled trait<->instance code from
creating circular dependencies that mess up tools.
(This is mostly a solution to the problem of "I have this trait that
gets instantiated N times, but I want a variable that only has 1
instance." And that problem also strongly affects unittesting!)

--- Subclasses ---
If there is a class named Kickback\Bazqux\Something
living in the file html/Kickback/Bazqux/Something.php,
then it is now possible to have auxiliary classes live
in the same file. They just have to be named things
like Kickback\Bazqux\Something__SomeIterator
or Kickback\Bazquz\Something__HelperThingie

Those too, can live in html/Kickback/Bazqux/Something.php.

The key is the double-underscore within the class name,
which is used to indicate that the thing is to be found
in the file represented by the class name at the left
of the double-underscore.

------------------

For now, all of the changes are limited to autoloading in the Kickback
namespace, so that they can't accidentally affect autoloading of
vendor/(3rd party) code. (Even though these names and tokens are fairly
unique and unlikely to do bad. Guarantees are nice, though.)