Add the ClassOfConstantIntegersTrait
PHP has `enum` types.
However, they are lacking in various ways:
* Don't support bitwise or arithmetic operations (so no flag sets!)
* Can't contain duplicate values (which happens sometimes, whether
    because an alias is helpful, or because some 3rd party code or API
    has duplicate values in their enumeration, and now we need to do
    the same thing to model their system state).
* Values must be literals or "literal expressions",
    meaning that you can't calculate constants from each other
    using arithmatic, ex:
    `case Count = self::Max + 1;` <- NOPE
    Note: PHP 8.2 may have lifted this limitation:
    ```
    The equivalent values may be a constant scalar expression.
    Prior to PHP 8.2.0, the equivalent values had to be literals or
    literal expressions. This means that constants and
    constant expressions were not supported.
    That is, 1 + 1 was allowed, but not 1 + SOME_CONST. is not.
    ```
* Having to use EnumName::Foo->value over and over to work
    around some of these limitations.

Source:
https://www.php.net/manual/en/language.enumerations.backed.php

Also good info:
https://stackoverflow.com/a/66208862

...

So, the usual alternative is classes-of-constant-integers:
```
final class ClassName
{
    use StaticClassTrait;

    public const FOO = 0;
    public const BAR = 1;
    public const BAZ = 2;
    public const RAB = 1; // Alias of BAR
    public const MAX = self::BAZ;
    public const COUNT = self::MAX+1;

    /** @param  self::*  $val */
    public static function stringize(int $val) : string
    {
        ...
    }
}
```

This is less restrictive, and with static analyzers like PHPStan,
we can still have type safety:
```
/**
* @param  ClassName::*  $fbb
*/
public static function do_a_thing(int $fbb) : void
{
    ...
}
```

The catch is that now we have to implement a bunch of boilerplate for
every class-of-integer-constants:
* Stringizer
* Conversion to array(s)
* Max/Min values to keep in sync, if needed
* And more.

This is a solvable problem, so this commit is the first iteration of
a solution for that problem.

This is useful for creating enumeration-like classes that house constant
values. These tend to be much more useful than the PHP enum type,
because the constants can be subjected to arithmetic operations
and generally behave... like integers.

Now it can look like this:
```
final class ClassName
{
    use StaticClassTrait;
    use ClassOfConstantIntegersTrait;

    public const FOO = 0;
    public const BAR = 1;
    public const BAZ = 2;
    public const RAB = 1; // Alias of BAR
    // For MAX, just call ClassName::max()
    // For COUNT, just call ClassName::count()

    // No need to implement stringizer. Or anything else.
}
```

Phew. So yeah. That exists now.

(And it's going to probably see a lot of use when handling database
flags and stuff. I might need to make a more specialized version for
handling bitwise-flags, because they are sparse when treated linearly,
but dense in log2. \mysqli already has a lot of constants and flags...)