Bugfix: init bootstrapper identified web vs cli incorrectly.
BEWARE: This commit will need to be applied in synchronous fashion to
both `beta` and `prod` systems!!

(It touches the more sensitive parts of the init bootstrapping system.
The parts that couldn't be abstracted out of synchronicity issues.
Darn.)

This commit also disables the website error handler _during CLI operation_
because it makes it very very difficult to understand CLI output.

(The error handler is still used in WEB context, of course.)

Anyways, this commit will probably be very important for everyone
to be able to run unittests from CLI without getting HTML vomit from the
error handler that was fooled into thinking it's running in web context
when it isn't.