Autoloader now ignores ALL '*_a' symbols, not just lowercase ones.
This is a tweak on the earlier recent autoloader commit that made the
autoloader ignore all symbols (in the Kickback namespace) that are
all-lowercase and end in `_a`.

Unfortunately, we have a bunch of symbols that are both "local aliases"
AND "not lowercase". This happens in some intractable situations too,
like when the symbol must contain a classname that it references.

So I am lifting the "lowercase" restriction and hoping that it will turn
out alright. (The `_a` alone is already very vanishingly rare,
so it should be fine.)