Better unittesting output, now with _some_ ANSI color \o/
This includes better logic for detecting when zend.assertions=1
isn't set: it'll try to figure it out at the beginning of the
test run, and abort in advance if it can prove that things
will just end up very misleading.

Also, if the zend.assertions error happens, it will tell you what value
it had (if it knows).

When using an intentional `assert(fail)` to test if assertions are live,
it will not longer print the distracting AssertionError output
and stack trace. It would seem that we can actually catch that exception
and get proof-positive without cluttering the terminal; it's great.

If the assertions aren't spicy, it will no longer throw an exception to
print its warning. It'll just print it with `echo`.

And while I can't do anything about the test blocks themselves just yet,
I at least made the headers/footers for the main unittest output
colorized and they should look much easier on the eyes.

...

This commit also comes with an array flattener for sequential arrays and
a (file) path normalizer (because PHP's `realpath` resolves symlinks and
may perform I/O, both of which can be quite undesirable at times).