Autoloader no longer tries to load the `scalar` type.
The `scalar` type is a typeclass used in at least PHPStan.

It is a useful shorthand for saying "all non-array things".

Of course, PHPStan invokes the autoloader on it for some yet undisclosed
reason, which leads to guaranteed autoloader error messages.

So just like the `never` type, we'll make a special bandage for this one
in the autoloader.

Now using `scalar` in PHPStan/PHPDoc comments will no longer generate
spurious and unhelpful error messages when running PHPStan.