Fix autoloading of subclasses.
So trying to autoload the (yet to be committed)
`\Kickback\Backend\Views\vDateTime__Tests` class,
present in `html/Kickback/Backend/Views/vDateTime.php`,
resulted in an autoload failure.

Turns out it was trimming the name from
`vDateTime__Tests` to `vDateTime__` and trying to load THAT.

This is fixed now, and properly trims
`vDateTime__Tests` to `vDateTime` and succeeds at loading.