Autoloader: optionally load IFooBar from FooBar.php
This is motivated by upcoming work on revamping the exception hierarchy.
Using subclass autoloading ... worked ... but it was pretty ugly in this
case. Notably, both the interface and its class counterpart are both
going to be very visible at some point in the usage pattern. Ex: the
exception class gets thrown, and the exception interface gets caught,
and if it doesn't, then that class will also appear in the debug output.
If the class has an ugly name like NotImplementedException__Thrown
then the stack trace looks more confusing. If we shift the burden to the
interface by naming the class NotImplementedException and the interface
NotImplementedException__Tag, then the catch-clauses can get verbose and
noisy. This allows us to use INotImplementedException instead in our
catch-clauses, which should be quite "good enough".

For anyone wondering: the interface-class duality works really well in
PHP because interfaces provide multiple inheritance, which allows for
exception ontologies that are tag-based and attributive, rather than
category-based and hierarchical. The hierarchies aren't good because you
end up with dilemmas like "do we organize file system exceptions by
Entry/Entity type or by (Failure Pattern)/Mode?" In such dilemmas, there
is no right answer, and you'll chop off one hand or the other. But with
tag-based models, you keep both hands, and can be very happy. Meanwhile,
the class part of the duality is necessary simply because interfaces
can't be instantiated, so there needs to be a class counterpart for the
code doing the throw statement to be able to instantiate.

Anyhow, that's how we got here.

Another thing that this autoloader change does is greatly simplify the
"avoid double-dipping" logic in the new subclass/trait/interface
features. It checks for declaration existance now instead of trying to
figure out when it needs to use `require_once`. This also gets rid of a
call to the `glob` function, which is file I/O. And it's always nice to
get rid of file I/O (optimization-wise). More work with tracing reveals
that we might not even need this logic at all, as the autoloader seems
to already be doing the same thing effectively on its own, which is
great (but probably undocumented? ah, oh well).