Str::is_ascii_identifier and Str::next_newline
Str::is_ascii_identifier is useful for parsing and
for data validation (because PHP doesn't seem to have this).
This is very general, and I'm surprised there didn't seem
to be an existing non-allocating (in principle) and
non-PCRE solution for this.

Str::next_newline is good for traversing a string line-by-line without
inducing spurious memory allocations. Not quite an immediate need, but
it was laying around from earlier work. I think its original motivation
was for creating text streams, loggers, or some kind of formatted output
that can handle useful things like formatting information into a margin
and indentating caller-supplied strings in uniform ways.