Add Meta::callable_to_unique_name and CategoriesOfCallables
This allows code to map callables to unique strings, which is useful for
identifying them. In this case, the use-case was identifying callables
for caching purposes, to avoid memory (re)allocations.
(Though the code depending on it is still a work-in-progress.)

On the downside, I'm not really happy with the CategoriesOfCallables
class being located in the `Primitives` namespace. It feels like this
should live somewhere else, but I'm not sure where to put it.

And I wonder if `Meta::callabale_to_unique_name` should really be
`Callable::to_unique_name` or something like that. Ah well, it's not
terribly important right now.