Autoloader improvements
Added syntax for autoloading subclasses with "numbered" designations.
Ex:
When autoloading `FunctionCapture_1s1c0`, it should check possibilities
that partially share the numbering, like `FunctionCapture__1s1.php`
or `FunctionCapture__1s.php`.

This commit also features a bugfix for autoloading eponymous interfaces,
which now allows it to handle something like `IO_IException`, whereas
previously it would fixate on the first 'I' instead of noticing the
ambiguity.

This work also involved implementing various string functions that are
now available in `Kickback\Common\Primitives\Str`. Some of the more
autoloader-specific name calculations are exposed in
`Kickback\Common\Primitives\Meta`. In either case, this allows the some
of the new functions to be unittested, which is something that isn't
doable from within the autoloader itself. But since the autoloader
essentially defines functions in its namespace, and is guaranteed to be
loaded by the time any class methods are called, it makes sense to create
OOP wrappers for these functions and then test their functionality.
This only applies to autoloader functions that don't do any IO and are
purely string analysis helpers. Further testing isn't impossible, but it
falls into the "high effort" category due to it requiring a lot of
mocking of I/O operations and possibly some PHP symbol handling
behaviors.

This commit also streamlines the handling of paths and argument passing
within the autoloader itself, which should make it easier to reason
about operations that "cut across" multiple paths, and also just makes
the function signatures look more reasonable.