Implement the `Kickback\PHPSTAN_RUNNING` constant.
This commit should be released to `beta` and `prod` simultaneously!
(It is in a very early part of initialization that is not protected
against document-root desynchronization.)

Details:

This constant is important for giving code the ability to selectively
exclude code from PHPStan analysis. This isn't normally something anyone
would want to do, but future commits make use of parametric boilerplate
that gets quite long and slow to analyse, while being so homogenous that
unlikely to be modified that the cost-to-benefit ratio just isn't there.

And this IS more effective than
```
if ( \Kickback\InitializationScripts\PARENT_PROCESS_TYPE === 'PHPSTAN' ) {
    return;
}
... not dead code according to PHPStan ...
... (but we want it to be) ...
```

... because PHP doesn't consider the above condition to create
"dead code", while this:
```
if ( \Kickback\PHPSTAN_RUNNING ) {
    return;
}
... IS dead code according to PHPStan ...
... allows us to exclude this code from analysis ...
```
... works perfectly.

(Perhaps the simplicity of a boolean constant makes it easier for
PHPStan to feel confident in its assertion. The logic for defining
`PARENT_PROCESS_TYPE` is also very non-trivial, while the logic for
defining `PHPSTAN_RUNNING` is nearly unconditional, only requiring
that it not be already defined. This is possible because
`PHPSTAN_RUNNING` is a much more specific test, from an init system
perspective, because it ONLY concerns PHPStan, and doesn't need to
perform other checks like Web-vs-CLI, some of which are very
complicated.)