Implement `KickbackException->say_before_message(...)`
(And also `KickbackException->say_after_message(...)`.)

This allows calling code to provide context for an exception without
needing to actually handle the exception.

It is VERY useful for layered systems where low-level components are not
aware of WHY an operation is happening, but DO have critical details.
Meanwhile higher-level components understand contextual information, but
do not know lower-level details.

This was somewhat tricky to implement due to PHP preventing the
`getMessage()` function of the `\Exception` class from being overridden,
and also because anything implementing `\Throwable` must also implement
`\Exception` (at least to be thrown/caught).

The way around the above is to realize that PHP _does_ allow
`\Exception->__toString()` to be overridden. This still leaves us with
the difficulty of figuring out how to format the resulting string in a
way that produces visually acceptable output when an exception isn't
caught and ends up printed somewhere.

The end result is a pretty nice `\Exception->__toString()`
implementation that can handle multiline messages elegantly, in addition
to the original goal of allowing different components to append/prepend
different contextual messages to the exception.