Various improvements/changes to ClassOfConstantIntegersTrait
The naming convention is hopefully easier to understand and less likely
to cause name collisions.

Notably, `stringize` is replaced with `name_of`, just because the method
name `stringize` is much more likely to be taken already by the class
that's using the trait.

This commit also adds the `nt_name_of` method which provides a
non-throwing way to retrieve constant names.