Add Kickback\Common\Meta\ZType
This class provides a way to store PHP types as integers instead
of the usual string returned by things like `\gettype` and
`\get_debug_type`.

Right now, it isn't used. But it is potentially handy for the WIP
unittesting framework or database abstraction layer.

Of course, an even more important reason for this commit is simply to
make this functionality available to anyone. Because it always seemed a
bit odd that it PHP primitive types were only provided as strings, never
integers, even though there is a small and finite number of them.

This means ZTypes are useful in a more general sense:
The list of type IDs is small and consecutive, so it can be used as an
array index any time some value needs to be looked up based on
the PHP type of another value.