Add Report(ing) class for multiple-exception awareness.
This also comes with a bunch of tweaks to things like the
`say_before_message` and `say_after_message` functions, and also
involved adding property-methods like `message()`, `file()`, and `line()`
to Kickback exceptions. These work around the limitation in PHP's
built-in exceptions that they don't allow us to reassing these
fields after an exception has been constructed. It is, however,
helpful to be able to change those after exception construction.
In the case of the `Report` class, this allows a more natural flow,
where the message describing the group of errors is in the "main"
exception message, at the end of the rest, and the caller would most
naturally assign this at the _end_ of Report generation.