Add `array::validate_*` functions; Also unify location parameters
This should make it easier to validate arrays that represent things like
JSON data, SQL results, config settings, etc. It won't cover all
use-cases, but this at least makes basic existence and type-agreement
checks more easy to do.

In particular, this makes it easier to do that validation _and create
good error reports when the validation fails_.

The methodology supports accumulating error messages, rather than just
reporting the first one encountered, which can reduce the number of
trial-and-error cycles in troubleshooting (and that's especially helpful
in situations where the cycle happens to be time-consuming or difficult
to reproduce, and it's not always clear when that will be the case).