Improve Reporting/Report->generate_exception
Previously, it was impractical to just throw the return value of this
function, as what would make sense in many cases.

That's because it returned a union type that included non-throwable
things like `MockException` and `null`.

Now it only returns a throwable object that also implements
`IKickbackThrowable`.

For cases where the possible `null` return value is desired
(e.g. to determine if the report had errors, for code that
doesn't already know), there is now a `generate_exception_if_errors`
method.

This also generally changed how exceptions are "typed" throughout the
class, as is necessary to ensure that they are throwable, but also
not constrained to one specific exception/throwable root class.