Helpers for KickbackThrowable->message() method.
Previously, you would have to _know_ that the exception is a
`KickbackThrowable` before being able to call `->message()`.
This is problematic if the exception _might_ be a `IKickbackThrowable`,
but might also be a native PHP `\Throwable`.

Although we will never have a `->message()` field on `\Throwable` and
have no way to inject that into the PHP runtime/Zend/etc, we do now
offer the `ThrowableOverrides::message($th)` function, which handles
the tedious logic like:
"if it's `IKickbackThrowable`, call `->message()`,
otherwise it's `\Throwable`, then call `->getMessage()".

The static class also implements
`->code()`, `->file()`, `->func()`, and `->line()`.

Whenever accessing those specific fields of an exception from within a
catch-clause (exception handler), we should always use the
`ThrowableOverrides` class to ensure that the correct information is
acquired. Simply using the default `->get*()` PHP methods may
result in stale or missing data in the case of Kickback Throwables
whenever the assignable alternatives were actually used.