Make `SERVICE_CREDENTIALS_SOURCES` constant public.
This is being done because it is nice for calling code to be able to
print a list of paths that are searched for the credentials file,
especially if there is an error with it. (And this is even more
practical now that KickbackException supports reasonable formatting of
multiline exception messages.)

This commit also adds some '\' (backslashes) to PHP built-in function
calls, as this is a more correct (and efficient) way to call these
globally-scoped symbols.